Unbreak RELEASEME_L10N_REQUIREMENT support
Dies otherwise with error message abozut non-existing dir in
FileUtils.rm_r(stat.dirs) for languages with zero translations
and whose dir had been removed during the zero percent check before.

+++ lib/releaseme/l10n.rb
@@ -132,7 +132,7 @@ module ReleaseMe
         completion_requirement = ENV['RELEASEME_L10N_REQUIREMENT'].to_i
         @statistics.languages.each do |stat|
           percent = stat.percent_translated
-          next if percent >= completion_requirement
+          next if percent >= completion_requirement || zero_percent_dropped.include?(stat.name)
           FileUtils.rm_r(stat.dirs)
           log_warn "#{stat.name} wasn't sufficiently translated #{percent}!"
         end
